pyproject.toml
[tool.poetry]
name = "hydromassnet"
version = "0.1.0"
description = "Bayesian Neural Networks to estimate neutral hydrogen mass in galaxies."
authors = ["Joelson Sartori Junior <joelsonsartori@gmail.com>"]
readme = "README.md"
license = "MIT"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.9,<3.13"
tensorflow = "^2.16.1"
tensorflow-probability = {version = "^0.24.0", extras = ["tf"]}
scikit-learn = "^1.5.0"
pandas = "^2.2.2"
matplotlib = "^3.9.0"
seaborn = "^0.13.2"
astroquery = "^0.4.7"
tqdm = "^4.66.4"
pyyaml = "^6.0.1"
lightgbm = "^4.6.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
